+++ src/renderer/features/drafts/index.ts
@@ -1,2 +1,3 @@
 export { DraftsSection } from './components/DraftsSection';
+export { type DraftSeed, createDraftModel } from './model/create-draft-model';
 export { type Draft } from '@/domains/backend';
