+++ src/renderer/shared/i18n/locales/en.json
@@ -1304,7 +1304,7 @@
       },
       "modalTitle": "Evidence & Salary",
       "promotionApplied": "Applied for promotion:",
-      "promotionAppliedAlertDescription": "You are already uploaded evidence for promotion. If you want to add it for retention, you can upload it again — the previous one will be replaced.",
+      "promotionAppliedAlertDescription": "You are already uploaded evidence for promotion. If you want to add it for retention, you can upload it again - the previous one will be replaced.",
       "promotionAppliedAlertTitle": "Promotion evidence is uploaded",
       "promotionEvidenceModalTitle": "Submit your promotion evidence",
       "promotionHelpRank1": "<li>Provide three examples of meaningful protocol contributions.</li><li>Actively design a deployable or future protocol component at PRP² standards.</li><li>Assist in analysis, formalization, or implementation at peer-reviewed or PRP standards.</li><li>Know Polkadot’s key goals and principles.</li><li>Examples: fixing key issues, enabling network fixes, proposing innovations, or improving protocol design.</li>",
@@ -2033,7 +2033,7 @@
         "invalidDataDescription": "File couldn’t be read. Please check data formatting and try again.",
         "invalidDataRow": "Row {row}",
         "invalidStructureDescription": "File couldn’t be read or contains invalid CSV structure. Check column headers and formatting, then try uploading again.",
-        "issueLine": "Row {row} — {path}: {message}",
+        "issueLine": "Row {row} - {path}: {message}",
         "issuesTitle": "Found {count} errors",
         "parsedFileDownloadDescription": "You can download parsed file to review them or open the preview. Please edit your file and upload it again.",
         "title": "CSV error",
@@ -2535,6 +2535,7 @@
       "overviewButton": "Overview",
       "editButton": "Edit",
       "submitButton": "Submit",
+      "addCallDataButton": "Add call data",
       "editDraftTitle": "Edit draft",
       "editSuccess": "Draft updated",
       "editError": "Failed to update draft",
@@ -2560,6 +2561,9 @@
       "stepMultisig": "Account",
       "stepCallData": "Call data",
       "stepConfirm": "Confirm",
+      "skipButton": "Skip",
+      "skipTooltip": "You'll be able to set the actual extrinsic when submitting this operation.",
+      "summaryNoCallData": "No call data - you'll add it when submitting this operation.",
       "linkCopied": "Draft link copied!",
       "shareDraftTooltip": "Share draft link with signatories",
       "selectMultisigForProxy": "Multisig for proxy",
@@ -3548,7 +3552,7 @@
     "renameHint": "Rename template",
     "renameTitle": "Rename template",
     "renameSave": "Save",
-    "staleTooltip": "Chain runtime was updated since this template was saved. It may fail to parse — re-save to refresh.",
+    "staleTooltip": "Chain runtime was updated since this template was saved. It may fail to parse - re-save to refresh.",
     "emptyListDescription": "Save an operation as a template to reuse its call data later.",
     "otherChains": "In other networks ({num})",
     "crossChainDisabled": "Switch network to apply",
